[SP-6750] [SP-6751] Backport of BISERVER-15167, BISERVER-15168: Schedules are lost during Backup-restore procedure even when restore is successful on same version of Pentaho Server (10.2 Suite) Restored schedules from previous versions are missing on latest scheduler-plugin (10.2 Suite)
Worked with @dcleao and @nawaz34-hitachivantara to address the fact that schedules were failing to import due to user roles/permissions, and output path validation for different importing user vs schedule user and vfs. So for this repo we:

    Made sure to be getting the user from the schedule request, not the session
    Use the schedule request user for output path validation via runAs

Refactored and updated UT coverage:
We are now testing on resolveOutputFilePathCore because the original method uses SecurityHelper.getInstance().runAsUser( getScheduleOwner(), this::resolveOutputFilePathCore ); and mocking the SecurityHelper runAsUser is unnecessary when we are ONLY testing for the same scheduler owner and session owner. runAs effectively does this and we are calling it because the person running the import/update could be different from the schedule owner.
[BACKLOG-43001] [BISERVER-15167] [BISERVER-15168] Restoring schedules from previous and current version

Move common logic into helper scheduleOutputPathResolver method because the following is true for almost all methods:

    session user is "admin"
    schedule owner is "admin"
    admin home folder is present and accessible
    common object instanciation and wiring

Provide overloaded method for cases where we want to

    set a jobName
    set a schedule owner other than "admin"
    change permissions/existence of "admin" home folder

Remove getUserHomeDirectoryPath() as this method will already return that value since we've set the "admin" session user

(cherry picked from commit 9a0e2f201452740b014c5cdb41e2244d4343c0aa)